Address comments #15
