+++ src/systems/systemDialog.component.tsx
@@ -13,10 +13,10 @@ import {
   Select,
   TextField,
 } from '@mui/material';
-import React from 'react';
-import { SystemImportanceType, SystemPost, useAddSystem } from '../api/systems';
 import { AxiosError } from 'axios';
-import { ErrorParsing } from '../app.types';
+import React from 'react';
+import { useAddSystem } from '../api/systems';
+import { ErrorParsing, SystemImportanceType, SystemPost } from '../app.types';
 
 export interface SystemDialogProps {
   open: boolean;
